pyproject.toml
[project]
name = "ai-agent-mvp"
version = "0.1.0"
description = "AI Agent MVP - Feishu URL to low-code workflow generation"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.2",
    "fastapi>=0.120.4",
    "langgraph>=1.0.2",
    "langgraph-cli[inmem]>=0.4.5",
    "langsmith>=0.4.39",
    "psutil>=7.1.3",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
